feat(feeds): add FEED_REFRESH periodic channel-refresh job
A self-paced virtual-thread loop refreshes all users_subscribed channels over a
configurable window (FEED_REFRESH on/off, FEED_REFRESH_MINUTES default 15), one
channel every period/count. Extracts channelResponse's side effects into shared
ChannelHelpers utils (videosTabInfo, federateChannelVideos, federateChannelInfo,
updateChannelVideos) used by both the handler and the loop.
