+++ src/main/java/me/kavin/piped/consts/Constants.java
@@ -68,6 +68,10 @@ public class Constants {
 
     public static final boolean DISABLE_PUBSUB;
 
+    public static final boolean FEED_REFRESH;
+
+    public static final int FEED_REFRESH_MINUTES;
+
     public static final String RYD_PROXY_URL;
 
     public static final List<String> SPONSORBLOCK_SERVERS;
@@ -147,6 +151,8 @@ public class Constants {
             FEED_RETENTION = Integer.parseInt(getProperty(prop, "FEED_RETENTION", "30"));
             DISABLE_TIMERS = Boolean.parseBoolean(getProperty(prop, "DISABLE_TIMERS", "false"));
             DISABLE_PUBSUB = Boolean.parseBoolean(getProperty(prop, "DISABLE_PUBSUB", "false"));
+            FEED_REFRESH = Boolean.parseBoolean(getProperty(prop, "FEED_REFRESH", "false"));
+            FEED_REFRESH_MINUTES = Integer.parseInt(getProperty(prop, "FEED_REFRESH_MINUTES", "15"));
             RYD_PROXY_URL = getProperty(prop, "RYD_PROXY_URL", "https://ryd-proxy.kavin.rocks");
             SPONSORBLOCK_SERVERS = List.of(getProperty(prop, "SPONSORBLOCK_SERVERS", "https://sponsor.ajay.app,https://sponsorblock.kavin.rocks")
                     .split(","));
