+++ src/main/java/me/kavin/piped/server/handlers/ChannelHandlers.java
@@ -41,106 +41,17 @@ public class ChannelHandlers {
 
         final ChannelInfo info = ChannelInfo.getInfo("https://youtube.com/" + channelPath);
 
-        final var preloadedVideosTab = collectPreloadedTabs(info.getTabs())
-                .stream()
-                .filter(tab -> tab.getContentFilters().contains(ChannelTabs.VIDEOS))
-                .findFirst();
-
-        final ChannelTabInfo tabInfo = preloadedVideosTab.isPresent() ?
-                ChannelTabInfo.getInfo(YOUTUBE_SERVICE, preloadedVideosTab.get()) :
-                null;
+        final ChannelTabInfo tabInfo = ChannelHelpers.videosTabInfo(info);
 
         final List<ContentItem> relatedStreams = tabInfo != null ? collectRelatedItems(tabInfo.getRelatedItems()) : List.of();
 
         if (tabInfo != null)
-            Multithreading.runAsync(() -> tabInfo.getRelatedItems()
-                    .stream().filter(StreamInfoItem.class::isInstance)
-                    .map(StreamInfoItem.class::cast)
-                    .forEach(infoItem -> {
-                        if (
-                                infoItem.getUploadDate() != null &&
-                                        System.currentTimeMillis() - infoItem.getUploadDate().offsetDateTime().toInstant().toEpochMilli()
-                                                < TimeUnit.DAYS.toMillis(Constants.FEED_RETENTION)
-                        )
-                            try {
-                                MatrixHelper.sendEvent("video.piped.stream.info", new FederatedVideoInfo(
-                                        StringUtils.substring(infoItem.getUrl(), -11), StringUtils.substring(infoItem.getUploaderUrl(), -24),
-                                        infoItem.getName(),
-                                        infoItem.getDuration(), infoItem.getViewCount())
-                                );
-                            } catch (IOException e) {
-                                throw new RuntimeException(e);
-                            }
-                    })
-            );
+            Multithreading.runAsync(() -> ChannelHelpers.federateChannelVideos(tabInfo));
 
-        Multithreading.runAsync(() -> {
-            try {
-                MatrixHelper.sendEvent("video.piped.channel.info", new FederatedChannelInfo(
-                        info.getId(), StringUtils.abbreviate(info.getName(), 100), info.getAvatars().isEmpty() ? null : info.getAvatars().getLast().getUrl(), info.isVerified())
-                );
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-        });
+        Multithreading.runAsync(() -> ChannelHelpers.federateChannelInfo(info));
 
         if (tabInfo != null)
-            Multithreading.runAsync(() -> {
-
-                var channel = DatabaseHelper.getChannelFromId(info.getId());
-
-                try (StatelessSession s = DatabaseSessionFactory.createStatelessSession()) {
-
-                    if (channel != null) {
-
-                        ChannelHelpers.updateChannel(s, channel, StringUtils.abbreviate(info.getName(), 100), info.getAvatars().isEmpty() ? null : info.getAvatars().getLast().getUrl(), info.isVerified());
-
-                        Set<String> ids = tabInfo.getRelatedItems()
-                                .stream()
-                                .filter(StreamInfoItem.class::isInstance)
-                                .map(StreamInfoItem.class::cast)
-                                .filter(item -> {
-                                    long time = item.getUploadDate() != null
-                                            ? item.getUploadDate().offsetDateTime().toInstant().toEpochMilli()
-                                            : System.currentTimeMillis();
-                                    return System.currentTimeMillis() - time < TimeUnit.DAYS.toMillis(Constants.FEED_RETENTION);
-                                })
-                                .map(item -> {
-                                    try {
-                                        return YOUTUBE_SERVICE.getStreamLHFactory().getId(item.getUrl());
-                                    } catch (ParsingException e) {
-                                        throw new RuntimeException(e);
-                                    }
-                                })
-                                .collect(Collectors.toUnmodifiableSet());
-
-                        List<Video> videos = DatabaseHelper.getVideosFromIds(s, ids);
-
-                        tabInfo.getRelatedItems()
-                                .stream()
-                                .filter(StreamInfoItem.class::isInstance)
-                                .map(StreamInfoItem.class::cast).forEach(item -> {
-                                    long time = item.getUploadDate() != null
-                                            ? item.getUploadDate().offsetDateTime().toInstant().toEpochMilli()
-                                            : System.currentTimeMillis();
-                                    if (System.currentTimeMillis() - time < TimeUnit.DAYS.toMillis(Constants.FEED_RETENTION))
-                                        try {
-                                            String id = YOUTUBE_SERVICE.getStreamLHFactory().getId(item.getUrl());
-                                            var video = videos.stream()
-                                                    .filter(v -> v.getId().equals(id))
-                                                    .findFirst();
-                                            if (video.isPresent()) {
-                                                VideoHelpers.updateVideo(id, item);
-                                            } else {
-                                                VideoHelpers.handleNewVideo("https://youtube.com/watch?v=" + id, time, channel);
-                                            }
-                                        } catch (Exception e) {
-                                            ExceptionHandler.handle(e);
-                                        }
-                                });
-                    }
-                }
-            });
+            Multithreading.runAsync(() -> ChannelHelpers.updateChannelVideos(info, tabInfo));
 
         String nextpage = null;
         if (tabInfo != null && tabInfo.hasNextPage()) {
