+++ src/main/java/me/kavin/piped/utils/ChannelHelpers.java
@@ -7,13 +7,22 @@ import com.rometools.modules.mediarss.types.PlayerReference;
 import com.rometools.modules.mediarss.types.Thumbnail;
 import com.rometools.rome.feed.synd.*;
 import me.kavin.piped.consts.Constants;
+import me.kavin.piped.utils.obj.MatrixHelper;
 import me.kavin.piped.utils.obj.db.Channel;
 import me.kavin.piped.utils.obj.db.Video;
+import me.kavin.piped.utils.obj.federation.FederatedChannelInfo;
+import me.kavin.piped.utils.obj.federation.FederatedVideoInfo;
 import okhttp3.Request;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.time.DurationFormatUtils;
 import org.apache.commons.text.StringEscapeUtils;
 import org.hibernate.StatelessSession;
+import org.schabi.newpipe.extractor.channel.ChannelInfo;
+import org.schabi.newpipe.extractor.channel.tabs.ChannelTabInfo;
+import org.schabi.newpipe.extractor.channel.tabs.ChannelTabs;
+import org.schabi.newpipe.extractor.exceptions.ExtractionException;
+import org.schabi.newpipe.extractor.exceptions.ParsingException;
+import org.schabi.newpipe.extractor.stream.StreamInfoItem;
 
 import java.io.IOException;
 import java.net.MalformedURLException;
@@ -22,7 +31,12 @@ import java.net.URL;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
+import java.util.Set;
+import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
 
+import static me.kavin.piped.consts.Constants.YOUTUBE_SERVICE;
+import static me.kavin.piped.utils.CollectionUtils.collectPreloadedTabs;
 import static me.kavin.piped.utils.URLUtils.rewriteURL;
 
 public class ChannelHelpers {
@@ -72,6 +86,103 @@ public class ChannelHelpers {
             var tr = s.beginTransaction();
             s.update(channel);
             tr.commit();
+        }
+    }
+
+    public static void updateChannelVideos(ChannelInfo info, ChannelTabInfo tabInfo) {
+
+        var channel = DatabaseHelper.getChannelFromId(info.getId());
+
+        try (StatelessSession s = DatabaseSessionFactory.createStatelessSession()) {
+
+            if (channel != null) {
+
+                updateChannel(s, channel, StringUtils.abbreviate(info.getName(), 100), info.getAvatars().isEmpty() ? null : info.getAvatars().getLast().getUrl(), info.isVerified());
+
+                Set<String> ids = tabInfo.getRelatedItems()
+                        .stream()
+                        .filter(StreamInfoItem.class::isInstance)
+                        .map(StreamInfoItem.class::cast)
+                        .filter(item -> {
+                            long time = item.getUploadDate() != null
+                                    ? item.getUploadDate().offsetDateTime().toInstant().toEpochMilli()
+                                    : System.currentTimeMillis();
+                            return System.currentTimeMillis() - time < TimeUnit.DAYS.toMillis(Constants.FEED_RETENTION);
+                        })
+                        .map(item -> {
+                            try {
+                                return YOUTUBE_SERVICE.getStreamLHFactory().getId(item.getUrl());
+                            } catch (ParsingException e) {
+                                throw new RuntimeException(e);
+                            }
+                        })
+                        .collect(Collectors.toUnmodifiableSet());
+
+                List<Video> videos = DatabaseHelper.getVideosFromIds(s, ids);
+
+                tabInfo.getRelatedItems()
+                        .stream()
+                        .filter(StreamInfoItem.class::isInstance)
+                        .map(StreamInfoItem.class::cast).forEach(item -> {
+                            long time = item.getUploadDate() != null
+                                    ? item.getUploadDate().offsetDateTime().toInstant().toEpochMilli()
+                                    : System.currentTimeMillis();
+                            if (System.currentTimeMillis() - time < TimeUnit.DAYS.toMillis(Constants.FEED_RETENTION))
+                                try {
+                                    String id = YOUTUBE_SERVICE.getStreamLHFactory().getId(item.getUrl());
+                                    var video = videos.stream()
+                                            .filter(v -> v.getId().equals(id))
+                                            .findFirst();
+                                    if (video.isPresent()) {
+                                        VideoHelpers.updateVideo(id, item);
+                                    } else {
+                                        VideoHelpers.handleNewVideo("https://youtube.com/watch?v=" + id, time, channel);
+                                    }
+                                } catch (Exception e) {
+                                    ExceptionHandler.handle(e);
+                                }
+                        });
+            }
+        }
+    }
+
+    public static ChannelTabInfo videosTabInfo(ChannelInfo info) throws ExtractionException, IOException {
+        var preloadedVideosTab = collectPreloadedTabs(info.getTabs())
+                .stream()
+                .filter(tab -> tab.getContentFilters().contains(ChannelTabs.VIDEOS))
+                .findFirst();
+        return preloadedVideosTab.isPresent() ? ChannelTabInfo.getInfo(YOUTUBE_SERVICE, preloadedVideosTab.get()) : null;
+    }
+
+    public static void federateChannelVideos(ChannelTabInfo tabInfo) {
+        tabInfo.getRelatedItems()
+                .stream().filter(StreamInfoItem.class::isInstance)
+                .map(StreamInfoItem.class::cast)
+                .forEach(infoItem -> {
+                    if (
+                            infoItem.getUploadDate() != null &&
+                                    System.currentTimeMillis() - infoItem.getUploadDate().offsetDateTime().toInstant().toEpochMilli()
+                                            < TimeUnit.DAYS.toMillis(Constants.FEED_RETENTION)
+                    )
+                        try {
+                            MatrixHelper.sendEvent("video.piped.stream.info", new FederatedVideoInfo(
+                                    StringUtils.substring(infoItem.getUrl(), -11), StringUtils.substring(infoItem.getUploaderUrl(), -24),
+                                    infoItem.getName(),
+                                    infoItem.getDuration(), infoItem.getViewCount())
+                            );
+                        } catch (IOException e) {
+                            throw new RuntimeException(e);
+                        }
+                });
+    }
+
+    public static void federateChannelInfo(ChannelInfo info) {
+        try {
+            MatrixHelper.sendEvent("video.piped.channel.info", new FederatedChannelInfo(
+                    info.getId(), StringUtils.abbreviate(info.getName(), 100), info.getAvatars().isEmpty() ? null : info.getAvatars().getLast().getUrl(), info.isVerified())
+            );
+        } catch (IOException e) {
+            throw new RuntimeException(e);
         }
     }
